Move-logic.publicresource-classes-to-ui.controller-#6354
* Create PublicResourcesServlet.java

* Update PublicResourcesServlet.java

* Create PublicImageServlet.java

* Delete PublicImageServlet.java

* Delete PublicResourcesServlet.java
